requirements.txt
 aiogram==2.25.1
python-dotenv==1.0.1
stripe==5.5.0
python-docx==0.8.11
pdfplumber==0.10.2
PyPDF2==3.0.1
requests==2.31.0
aiohttp==3.8.5
fpdf2==2.7.5
aiofiles==23.2.1
python -m pip install -r requirements.txt